# sdks/sdkwork-voice-backend-sdk/openapi/sdkwork-voice-backend-api.flutter.sdkgen.yaml
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Voice Backend API",
    "version": "1.0.0",
    "description": "Backend/admin contract for SDKWork Voice provider routes, generation tasks, webhook reconciliation, request logs, and audio artifacts.",
    "x-sdkwork-api-authority": "sdkwork-voice-backend-api",
    "x-sdkwork-sdk-family": "sdkwork-voice-backend-sdk",
    "x-sdkwork-audience": "Backend consoles, operators, control-plane integrations, and admin automation"
  },
  "servers": [
    {
      "url": "http://localhost:18096",
      "description": "Local sdkwork-voice runtime"
    }
  ],
...
  "security": [
    {
      "AuthToken": [],
      "AccessToken": []
    }
  ],
  "paths": {
...
    "/backend/v3/api/voice/audio_artifacts/{audioArtifactId}": {
      "delete": {
        "tags": [
          "voice"
        ],
        "summary": "Audio Artifacts delete.",
        "operationId": "audioArtifacts.delete",
        "parameters": [
          {
            "name": "audioArtifactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkResourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-voice",
        "x-sdkwork-api-authority": "sdkwork-voice-backend-api",
        "x-sdkwork-domain": "voice",
        "x-sdkwork-resource": "audioArtifacts",
        "x-sdkwork-request-context": "AppRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-source": "<sdkwork-voice>/crates/sdkwork-routes-voice-backend-api/src/manifest.rs",
        "x-sdkwork-source-route-crate": "sdkwork-routes-voice-backend-api",
        "x-sdkwork-auth-mode": "dual-token"
      },
      "get": {
        "tags": [
          "voice"
        ],
        "summary": "Audio Artifacts retrieve.",
        "operationId": "audioArtifacts.retrieve",
        "parameters": [
          {
            "name": "audioArtifactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkResourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-voice",
        "x-sdkwork-api-authority": "sdkwork-voice-backend-api",
        "x-sdkwork-domain": "voice",
        "x-sdkwork-resource": "audioArtifacts",
        "x-sdkwork-request-context": "AppRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-source": "<sdkwork-voice>/crates/sdkwork-routes-voice-backend-api/src/manifest.rs",
        "x-sdkwork-source-route-crate": "sdkwork-routes-voice-backend-api",
        "x-sdkwork-auth-mode": "dual-token"
      }
    },
...
    "/backend/v3/api/voice/provider_routes/{providerRouteId}": {
      "delete": {
        "tags": [
          "voice"
        ],
        "summary": "Provider Routes delete.",
        "operationId": "providerRoutes.delete",
        "parameters": [
          {
            "name": "providerRouteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkResourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-voice",
        "x-sdkwork-api-authority": "sdkwork-voice-backend-api",
        "x-sdkwork-domain": "voice",
        "x-sdkwork-resource": "providerRoutes",
        "x-sdkwork-request-context": "AppRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-source": "<sdkwork-voice>/crates/sdkwork-routes-voice-backend-api/src/manifest.rs",
        "x-sdkwork-source-route-crate": "sdkwork-routes-voice-backend-api",
        "x-sdkwork-auth-mode": "dual-token"
      },
      "get": {
        "tags": [
          "voice"
        ],
        "summary": "Provider Routes retrieve.",
        "operationId": "providerRoutes.retrieve",
        "parameters": [
          {
            "name": "providerRouteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkResourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-voice",
        "x-sdkwork-api-authority": "sdkwork-voice-backend-api",
        "x-sdkwork-domain": "voice",
        "x-sdkwork-resource": "providerRoutes",
        "x-sdkwork-request-context": "AppRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-source": "<sdkwork-voice>/crates/sdkwork-routes-voice-backend-api/src/manifest.rs",
        "x-sdkwork-source-route-crate": "sdkwork-routes-voice-backend-api",
        "x-sdkwork-auth-mode": "dual-token"
      },
      "patch": {
        "tags": [
          "voice"
        ],
        "summary": "Provider Routes update.",
        "operationId": "providerRoutes.update",
        "parameters": [
          {
            "name": "providerRouteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkResourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-voice",
        "x-sdkwork-api-authority": "sdkwork-voice-backend-api",
        "x-sdkwork-domain": "voice",
        "x-sdkwork-resource": "providerRoutes",
        "x-sdkwork-request-context": "AppRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-source": "<sdkwork-voice>/crates/sdkwork-routes-voice-backend-api/src/manifest.rs",
        "x-sdkwork-source-route-crate": "sdkwork-routes-voice-backend-api",
        "x-sdkwork-auth-mode": "dual-token",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceOperationCommand"
              }
            }
          }
        }
      }
    },
...
    "/backend/v3/api/voice/tasks/{taskId}": {
      "get": {
        "tags": [
          "voice"
        ],
        "summary": "Tasks retrieve.",
        "operationId": "tasks.retrieve",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkResourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-voice",
        "x-sdkwork-api-authority": "sdkwork-voice-backend-api",
        "x-sdkwork-domain": "voice",
        "x-sdkwork-resource": "tasks",
        "x-sdkwork-request-context": "AppRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-source": "<sdkwork-voice>/crates/sdkwork-routes-voice-backend-api/src/manifest.rs",
        "x-sdkwork-source-route-crate": "sdkwork-routes-voice-backend-api",
        "x-sdkwork-auth-mode": "dual-token"
      }
    },
...
    "/backend/v3/api/voice/tasks/{taskId}/reconcile": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "Tasks reconcile.",
        "operationId": "tasks.reconcile",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkCommandResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-voice",
        "x-sdkwork-api-authority": "sdkwork-voice-backend-api",
        "x-sdkwork-domain": "voice",
        "x-sdkwork-resource": "tasks",
        "x-sdkwork-request-context": "AppRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-source": "<sdkwork-voice>/crates/sdkwork-routes-voice-backend-api/src/manifest.rs",
        "x-sdkwork-source-route-crate": "sdkwork-routes-voice-backend-api",
        "x-sdkwork-auth-mode": "dual-token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceTaskReconcileCommand"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "SDKWork auth token carried as Authorization: Bearer <auth_token>."
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token",
        "description": "SDKWork access isolation token."
      }
    },
    "schemas": {
      "SdkWorkApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ],
            "default": 0,
            "minimum": 0,
            "maximum": 0
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
...
      "SdkWorkPlatformErrorCode": {
        "type": "integer",
        "format": "int32",
        "minimum": 40001,
        "maximum": 79999,
        "description": "Platform or domain error code per API_SPEC.md §15.3."
      },
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/SdkWorkPlatformErrorCode"
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
...
      "SdkWorkResourceResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/SdkWorkResourceData"
              }
            }
          }
        ]
      },
...
      "SdkWorkCommandResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/SdkWorkCommandData"
              }
            }
          }
        ]
      },
      "VoiceOperationCommand": {
        "type": "object",
        "additionalProperties": true,
        "description": "Operation-specific command payload for SDKWork Voice administrative operations and provider extension commands.",
        "properties": {
          "mediaResource": {
            "$ref": "#/components/schemas/MediaResource"
          }
        }
      },
...
    }
  },
...
}